+++ apps/wsdjs_web/lib/wsdjs_web/views/user_view.ex
@@ -1,7 +1,13 @@
 defmodule WsdjsWeb.UserView do
   use WsdjsWeb, :view
 
+  alias Wsdjs.Accounts.User
+
   def can_edit_page?(user, current_user) do
     current_user == user
   end
+
+  def name_or_email(%User{email: email, name: nil, djname: nil}), do: email
+  def name_or_email(%User{name: name}) when is_binary(name), do: name
+  def name_or_email(%User{djname: djname}) when is_binary(djname), do: djname
 end